refactor: RacingHistories 에 우승자 계산 로직 추가
