+++ src/main/kotlin/racingcar/model/Car.kt
@@ -12,7 +12,9 @@ data class Car(val name: String, private var _position: Int = 0) {
     }
 
     companion object {
-        private const val CAR_NAME_VALIDATION_MESSAGE = "자동차 이름은 5자를 초과할 수 없습니다.";
+        private const val CAR_NAME_VALIDATION_MESSAGE = "자동차 이름은 5자를 초과할 수 없습니다."
         private const val CAR_NAME_MAX_BOUND = 5
     }
 }
+
+fun createCars(namesOfCars: List<String>): List<Car> = namesOfCars.map { Car(name = it) }
