--- src/main/kotlin/racingcar/model/CarInitializer.kt
@@ -1,5 +1,0 @@
-package racingcar.model
-
-object CarInitializer {
-    fun createCars(namesOfCars: List<String>): List<Car> = namesOfCars.map { Car(name = it) }
-}
